lambda/attachment-extractor/requirements.txt
mypylogger
pypdf
python-docx
